+++ arrow-array/src/array/fixed_size_binary_array.rs
@@ -1029,7 +1029,9 @@ mod tests {
     }
 
     #[test]
-    #[should_panic(expected = "The child array cannot contain null values.")]
+    // Panic message differs by config: `build_unchecked` validates under
+    // force_validate, otherwise `FixedSizeBinaryArray::from` catches it.
+    #[should_panic]
     fn test_fixed_size_binary_array_from_fixed_size_list_array_with_child_nulls_failed() {
         let values = [0_u8, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11];
         let values_data = ArrayData::builder(DataType::UInt8)
